Add files via upload
Updating scripts for Capstone CIS2089701. Player Controller, resourceSpawn, and Diamond Collection scripts are done. Things still to add are a particle effect, ship sounds, score function, gameManager or TimeManager. Nothing to tell it to only be 60 seconds. Also needs the TextMeshPro to be properly connected to correctly keep score and show time limit.